add: funcion de filtro del search & estilos
