+++ src/App.js
@@ -1,6 +1,7 @@
 import React, { createContext } from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { Home, Header } from './components/';
+import { Searchbar } from './components/searchbar/Searchbar';
 import { useGeolocation } from './hooks&aux/useGeolocation';
 
 export const LocationContext = createContext()
@@ -12,6 +13,7 @@ export const App = () => {
     <Router>
       <LocationContext.Provider value={currentLocation}>
         <Header />
+        <Searchbar />
         <Routes>
           <Route path='/' element={<Home />} />
           <Route path='*' element={<h1>404 Not Found</h1>} />
