+++ src/components/searchbar/searchbar.module.css
@@ -1,0 +1,15 @@
+.inputContainer {
+    margin: 16px;
+}
+
+.inputContainer input {
+    box-sizing: border-box;
+    width: 343px;
+    height: 41px;
+    padding: 13px 12px;
+    border-radius: 10px;
+    border: none;
+    background-color: #F1F2F6;
+    font-family: inherit;
+    font-size: 1.2rem;
+}
